requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
unstructured[all-docs]==0.11.2
marker-pdf==0.2.4
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
scrubadub==2.0.1
spacy==3.7.2
torch==2.1.1
transformers==4.35.2
detectron2==0.6
layoutparser==0.3.4
Pillow==10.1.0
python-magic==0.4.27
pypdf==3.17.1
pdfplumber==0.10.3
pytesseract==0.3.10
opencv-python==4.8.1.78
